eslint fix for server

+++ server/helpers/file.js
@@ -7,29 +7,29 @@ const fileTemplate = require('../templates/cron-file');
 
 const { ENV } = process.env;
 
-const dir = ENV === 'DOCKER' ? `/var/lib/cronker/crons` : __dirname+`/../../crons`;
-const filePath = (cron) =>  ENV === 'DOCKER' ? `/var/lib/cronker/crons/${cron.id}.js` : __dirname+`/../../crons/${cron.id}.js`;
+const dir = ENV === 'DOCKER' ? '/var/lib/cronker/crons' : `${__dirname}/../../crons`;
+const getFilePath = (cron) => (ENV === 'DOCKER' ? `/var/lib/cronker/crons/${cron.id}.js` : `${__dirname}/../../crons/${cron.id}.js`);
 
 exports.createCrons = () => {
-  if (!fs.existsSync(dir)){
+  if (!fs.existsSync(dir)) {
     fs.mkdirSync(dir);
   }
-}
+};
 
 exports.createFileForCron = (cron) => {
+  const filePath = getFilePath(cron);
   const fileData = fileTemplate(cron);
-  if (!fs.existsSync(filePath))
-    fs.openSync(filePath, 'w');
+  if (!fs.existsSync(filePath)) fs.openSync(filePath, 'w');
 
   fs.writeFileSync(filePath, fileData, 'utf-8', {
     flag: 'wx'
   });
-}
+};
 
 exports.updateFileForCron = (cron) => {
+  const filePath = getFilePath(cron);
   const fileData = fileTemplate(cron);
-  if (fs.existsSync(filePath))
-    fs.unlinkSync(filePath);
+  if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
 
   fs.writeFileSync(filePath, fileData, 'utf-8', {
     flag: 'wx'
@@ -37,11 +37,12 @@ exports.updateFileForCron = (cron) => {
 
   require('./cron').requireAll();
   setTimeout(() => pm2.restart('index'), 500);
-}
+};
 
 exports.deleteFileForCron = async (id) => {
+  const filePath = getFilePath({ id });
   if (fs.existsSync(filePath)) {
     fs.unlinkSync(filePath);
     setTimeout(() => pm2.restart('index'), 500);
   }
-}
+};
